Pull log & pow wrapper functions into separate file
