+++ rpm/dnf/glib_log_pow_wrapper.go
@@ -1,0 +1,8 @@
+// +build dnf
+// +build tests
+
+package dnf
+
+// #cgo LDFLAGS: -Wl,--wrap=log -Wl,--wrap=pow
+// #include "glib_log_pow_wrapper.h"
+import "C"
